pyproject.toml
[tool.poetry]
package-mode = true
name = "lfprop"
version = "0.1.0"
description = "A package for gradient-free neural network training using LFP"
readme = "README.md"
authors = [
    "Leander Weber <leander.weber@hhi.fraunhofer.de>",
]
maintainers = [
    "Leander Weber <leander.weber@hhi.fraunhofer.de>",
]
keywords = ["explainable ai", "xai", "machine learning", "deep learning", "gradient-free optimization"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
license = "BSD-3-Clause"
exclude = [
    "cluster_run",
    "configs",
    "experiment_utils",
    "minimal-example-data",
    "wandb",
    "nbs",
    ".gitignore",
    "minimal_example.ipynb",
    "run_experiments.py"
]

[tool.poetry.dependencies]
python = ">=3.11"
zennit = "^0.5.1"
lxt = "^0.6.1"
wandb = { version = "^0.19.4", optional = true }
notebook = { version = "^7.3.2", optional = true }
joblib = { version = "^1.4.2", optional = true }
pandas = { version = "^2.2.3", optional = true }
scikit-learn = { version = "^1.6.1", optional = true }
matplotlib = { version = "<3.8", optional = true }
torcheval = { version = "^0.0.7", optional = true }

[tool.poetry.extras]
quickstart = ["notebook", "torcheval"]
full = ["wandb", "notebook", "joblib", "pandas", "scikit-learn", "matplotlib", "torcheval"]

[project]
license = { file = "LICENSE" }

[project.urls]
Source = "https://github.com/leanderweber/layerwise-feedback-propagation"
Issues = "https://github.com/leanderweber/layerwise-feedback-propagation/issues"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
